requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy>=1.21.0
supabase==1.0.3
python-dotenv==1.0.0
pillow==10.0.1
pydantic==2.4.2
requests==2.31.0
face_recognition==1.3.0
dlib==19.24.6